+++ data-raw/allowed_inputs_outputs.R
@@ -52,7 +52,8 @@ data("priors", package = "dynwrap")
 allowed_inputs <- tribble(
   ~input_id, ~description,
   "expression", "Expression matrix (sparse)",
-  "counts", "Raw counts matrix (sparse)"
+  "counts", "Raw counts matrix (sparse)",
+  "expression_projected", "Projected expression matrix based on RNA velocity (sparse)"
 ) %>% bind_rows(
   priors %>% select(input_id = prior_id, description = description)
 )