+++ inst/NEWS.md
@@ -1,3 +1,15 @@
+# dynwrap 1.1 (07-05-2019)
+
+* FEATURE: RNA velocity data can now be included in the wrapper
+
+* FEATURE: RNA velocity projected expression can now be given to methods
+
+* FEATURE: Added `orient_topology_to_velocity()` to orient the edges of a trajectory based on the velocity vectors
+
+* FEATURE: `calculate_geodesic_distances()` now has a `directed` parameter, with which you can calculate directed geodesic distances. Unreachable cells will receive an infinite distance
+
+* FEATURE: Added `projected_trajectory()` (ported from dynplot). `add_dimred()` will now by default also add a projection of the trajectory to the data
+
 # dynwrap 1.0.1 (09-04-2019)
 
 * MINOR CHANGE: `infer_trajectory(..., debug = TRUE)` will now return commands for both entering the container and running the `dyncli::main()` function inside the container